# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ClickEvent": {
        "description": "Click event data.",
        "properties": {
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "source_page": {
            "title": "Source Page",
            "type": "string"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "required": [
          "event_type",
          "source_page"
        ],
        "title": "ClickEvent",
        "type": "object"
      },
      "ClickResponse": {
        "description": "Response for click tracking.",
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "total_clicks": {
            "title": "Total Clicks",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "total_clicks"
        ],
        "title": "ClickResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LegacyAccountMergeRequest": {
        "properties": {
          "legacy_user_id": {
            "format": "uuid",
            "title": "Legacy User Id",
            "type": "string"
          },
          "proof_id": {
            "format": "uuid",
            "title": "Proof Id",
            "type": "string"
          }
        },
        "required": [
          "legacy_user_id",
          "proof_id"
        ],
        "title": "LegacyAccountMergeRequest",
        "type": "object"
      },
      "LegacyAccountMergeResponse": {
        "properties": {
          "canonical_user_id": {
            "format": "uuid",
            "title": "Canonical User Id",
            "type": "string"
          },
          "legacy_user_id": {
            "format": "uuid",
            "title": "Legacy User Id",
            "type": "string"
          },
          "merge_id": {
            "format": "uuid",
            "title": "Merge Id",
            "type": "string"
          }
        },
        "required": [
          "merge_id",
          "legacy_user_id",
          "canonical_user_id"
        ],
        "title": "LegacyAccountMergeResponse",
        "type": "object"
      },
...
      "PaymentClaimProofRequest": {
        "properties": {
          "legacy_user_id": {
            "format": "uuid",
            "title": "Legacy User Id",
            "type": "string"
          },
          "payment_reference": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Payment Reference",
            "type": "string"
          }
        },
        "required": [
          "legacy_user_id",
          "payment_reference"
        ],
        "title": "PaymentClaimProofRequest",
        "type": "object"
      },
      "PaymentClaimProofResponse": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "proof_id": {
            "format": "uuid",
            "title": "Proof Id",
            "type": "string"
          }
        },
        "required": [
          "proof_id",
          "expires_at"
        ],
        "title": "PaymentClaimProofResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API for AI Wedding Photo generation using InstantID",
    "title": "AI Wedding Photo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/click": {
      "post": {
        "description": "Track ad/banner click events.\n\nStores daily aggregated counts in DB with a best-effort in-memory fallback.",
        "operationId": "track_click_api_v1_analytics_click_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClickEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClickResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Track Click",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/stats": {
      "get": {
        "description": "Get analytics statistics (admin only in production).",
        "operationId": "get_stats_api_v1_analytics_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stats",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/auth/account-claims/merge": {
      "post": {
        "operationId": "merge_empty_legacy_account_api_v1_auth_account_claims_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyAccountMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegacyAccountMergeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Merge Empty Legacy Account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/account-claims/payment-proof": {
      "post": {
        "operationId": "create_payment_claim_proof_api_v1_auth_account_claims_payment_proof_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentClaimProofRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentClaimProofResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Payment Claim Proof",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}